move prep logic to a class
