+++ .gitignore
@@ -40,3 +40,6 @@ override.tf.json
 # Ignore CLI configuration files
 .terraformrc
 terraform.rc
+
+# Ignore PyCharm
+.idea
