+++ data/prep/clubs.csv
@@ -1,0 +1,242 @@
+club_id,name,domestic_competition
+27,fc-bayern-munchen,L1
+16,borussia-dortmund,L1
+23826,rasenballsport-leipzig,L1
+15,bayer-04-leverkusen,L1
+18,borussia-monchengladbach,L1
+533,tsg-1899-hoffenheim,L1
+44,hertha-bsc,L1
+82,vfl-wolfsburg,L1
+24,eintracht-frankfurt,L1
+79,vfb-stuttgart,L1
+60,sc-freiburg,L1
+33,fc-schalke-04,L1
+3,1-fc-koln,L1
+167,fc-augsburg,L1
+86,sv-werder-bremen,L1
+39,1-fsv-mainz-05,L1
+89,1-fc-union-berlin,L1
+10,arminia-bielefeld,L1
+660,shakhtar-donetsk,UKR1
+338,dynamo-kiew,UKR1
+10690,zorya-lugansk,UKR1
+2740,vorskla-poltava,UKR1
+16239,desna-chernigiv,UKR1
+48332,kolos-kovalivka,UKR1
+18303,pfk-oleksandria,UKR1
+6993,ilyichivets-mariupol,UKR1
+60551,sk-dnipro-1,UKR1
+23611,olimpik-donetsk,UKR1
+31,fc-liverpool,GB1
+18105,fk-lviv,UKR1
+48726,rukh-vinnyky,UKR1
+49364,ingulets-pyatykhatky,UKR1
+61825,fk-minaj,UKR1
+281,manchester-city,GB1
+631,fc-chelsea,GB1
+985,manchester-united,GB1
+148,tottenham-hotspur,GB1
+11,fc-arsenal,GB1
+29,fc-everton,GB1
+1003,leicester-city,GB1
+543,wolverhampton-wanderers,GB1
+405,aston-villa,GB1
+762,newcastle-united,GB1
+379,west-ham-united,GB1
+1237,brighton-amp-hove-albion,GB1
+180,fc-southampton,GB1
+873,crystal-palace,GB1
+931,fc-fulham,GB1
+399,leeds-united,GB1
+2282,fc-brugge,BE1
+294,benfica-lissabon,PO1
+350,sheffield-united,GB1
+1132,fc-burnley,GB1
+984,west-bromwich-albion,GB1
+131,fc-barcelona,ES1
+1184,krc-genk,BE1
+58,rsc-anderlecht,BE1
+3057,standard-luttich,BE1
+157,kaa-gent,BE1
+1096,royal-fc-antwerpen,BE1
+172,rsc-charleroi,BE1
+520,cercle-brugge,BE1
+2727,oud-heverlee-leuven,BE1
+354,kv-mechelen,BE1
+41274,k-fc-o-beerschot-wilrijk,BE1
+29228,mouscron-peruwelz,BE1
+3508,sv-zulte-waregem,BE1
+601,kv-kortrijk,BE1
+475,vv-st-truiden,BE1
+1245,kas-eupen,BE1
+2861,kv-oostende,BE1
+28643,waasland-beveren,BE1
+964,zenit-st-petersburg,RU1
+583,fc-paris-saint-germain,FR1
+720,fc-porto,PO1
+336,sporting-lissabon,PO1
+1075,sc-braga,PO1
+2420,vitoria-guimaraes-sc,PO1
+3329,fc-famalicao,PO1
+2503,boavista-porto-fc,PO1
+2425,rio-ave-fc,PO1
+2995,fc-pacos-de-ferreira,PO1
+1301,cs-maritimo,PO1
+4294,sc-farense,PO1
+7378,portimonense-sc,PO1
+2423,cd-santa-clara,PO1
+979,moreirense-fc,PO1
+2457,cf-belenenses-lissabon,PO1
+982,cd-nacional,PO1
+2424,gil-vicente-fc,PO1
+7179,cd-tondela,PO1
+610,ajax-amsterdam,NL1
+13,atletico-madrid,ES1
+418,real-madrid,ES1
+368,fc-sevilla,ES1
+681,real-sociedad-san-sebastian,ES1
+1049,fc-valencia,ES1
+1050,fc-villarreal,ES1
+621,athletic-bilbao,ES1
+150,real-betis-sevilla,ES1
+3709,fc-getafe,ES1
+940,celta-vigo,ES1
+16795,fc-granada,ES1
+3368,ud-levante,ES1
+331,ca-osasuna,ES1
+1108,deportivo-alaves,ES1
+1533,sd-eibar,ES1
+2410,zska-moskau,RU1
+16704,fk-krasnodar,RU1
+232,spartak-moskau,RU1
+932,lokomotiv-moskau,RU1
+121,dinamo-moskau,RU1
+2698,rubin-kazan,RU1
+1083,fk-rostov,RU1
+3725,terek-grozny,RU1
+41231,fk-sochi,RU1
+3729,arsenal-tula,RU1
+11127,ural-sverdlovskaya-oblast,RU1
+3719,fk-khimki,RU1
+3609,rotor-volgograd,RU1
+28095,fk-ufa,RU1
+41201,fk-tambov,RU1
+1041,olympique-lyon,FR1
+162,as-monaco,FR1
+1082,osc-lille,FR1
+273,fc-stade-rennes,FR1
+244,olympique-marseille,FR1
+417,ogc-nizza,FR1
+667,racing-strassburg,FR1
+969,hsc-montpellier,FR1
+995,fc-nantes,FR1
+618,as-saint-etienne,FR1
+1421,stade-reims,FR1
+40,fc-girondins-bordeaux,FR1
+3911,stade-brest-29,FR1
+1420,sco-angers,FR1
+826,rc-lens,FR1
+383,psv-eindhoven,NL1
+1090,az-alkmaar,NL1
+234,feyenoord-rotterdam,NL1
+200,fc-utrecht,NL1
+306,sc-heerenveen,NL1
+403,willem-ii-tilburg,NL1
+499,vitesse-arnheim,NL1
+202,fc-groningen,NL1
+317,fc-twente-enschede,NL1
+1268,ado-den-haag,NL1
+1269,pec-zwolle,NL1
+1304,heracles-almelo,NL1
+385,fortuna-sittard,NL1
+468,sparta-rotterdam,NL1
+1283,fc-emmen,NL1
+1426,vvv-venlo,NL1
+235,rkc-waalwijk,NL1
+366,real-valladolid,ES1
+5358,sd-huesca,ES1
+2687,cadiz-cf,ES1
+1531,fc-elche,ES1
+371,celtic-glasgow,SC1
+347,fc-metz,FR1
+2969,fco-dijon,FR1
+1158,fc-lorient,FR1
+1160,olympique-nimes,FR1
+865,fc-midtjylland,DK1
+124,glasgow-rangers,SC1
+903,hibernian-fc,SC1
+370,aberdeen-fc,SC1
+987,motherwell-fc,SC1
+1519,dundee-united-fc,SC1
+1241,livingston-fc,SC1
+2553,kilmarnock-fc,SC1
+2578,st-johnstone-fc,SC1
+2759,ross-county-fc,SC1
+2999,hamilton-academical-fc,SC1
+465,st-mirren-fc,SC1
+190,fc-kopenhagen,DK1
+206,brondby-if,DK1
+2778,fc-nordsjaelland,DK1
+678,aarhus-gf,DK1
+1053,aalborg-bk,DK1
+173,odense-boldklub,DK1
+443,vejle-boldklub,DK1
+5817,sonderjyske,DK1
+5724,randers-fc,DK1
+369,lyngby-bk,DK1
+2414,ac-horsens,DK1
+141,galatasaray-istanbul,TR1
+36,fenerbahce-istanbul,TR1
+449,trabzonspor,TR1
+6890,istanbul-buyuksehir-belediyespor,TR1
+114,besiktas-istanbul,TR1
+10484,kasimpasa,TR1
+11282,alanyaspor,TR1
+126,caykur-rizespor,TR1
+2381,sivasspor,TR1
+1467,goztepe-izmir,TR1
+2832,gaziantep-buyuksehir-belediyespor,TR1
+6646,localtime-fatih-karagumruk,TR1
+2293,torku-konyaspor,TR1
+868,mke-ankaragucu,TR1
+19789,yeni-malatyaspor,TR1
+7775,hatayspor,TR1
+3205,kayserispor,TR1
+589,medical-park-antalyaspor,TR1
+833,tekden-denizlispor,TR1
+820,genclerbirligi-ankara,TR1
+39722,erzurum-buyuksehir-belediyespor,TR1
+683,olympiakos-piraus,GR1
+1091,paok-thessaloniki,GR1
+2441,aek-athen,GR1
+265,panathinaikos-athen,GR1
+605,aris-thessaloniki,GR1
+653,ofi-kreta,GR1
+3060,atromitos-athen,GR1
+4603,ae-larisa,GR1
+6418,panetolikos-agrinio,GR1
+6676,asteras-tripolis,GR1
+3999,pas-lamia-1964,GR1
+60949,volos-nfc,GR1
+2671,pas-giannina,GR1
+441,apollon-smyrnis,GR1
+506,juventus-turin,IT1
+46,inter-mailand,IT1
+6195,ssc-neapel,IT1
+5,ac-mailand,IT1
+12,as-rom,IT1
+800,atalanta-bergamo,IT1
+398,lazio-rom,IT1
+6574,us-sassuolo,IT1
+430,ac-florenz,IT1
+1390,cagliari-calcio,IT1
+416,fc-turin,IT1
+276,hellas-verona,IT1
+1025,fc-bologna,IT1
+130,fc-parma,IT1
+410,udinese-calcio,IT1
+1038,uc-sampdoria,IT1
+252,fc-genua-1893,IT1
+4171,benevento-calcio,IT1
+3522,spezia-calcio,IT1
+4083,fc-crotone,IT1
